Extracted logic for building reservation confirmation email to a separate class
