+++ hotel-booking-app-service/src/test/java/com/application/hotelbooking/services/ReservationConfirmationEmailServiceTest.java
@@ -1,0 +1,4 @@
+package com.application.hotelbooking.services;
+
+public class ReservationConfirmationEmailServiceTest {
+}
