+++ hotel-booking-app-service/src/test/java/com/application/hotelbooking/services/ReservationServiceImplTest.java
@@ -60,10 +60,7 @@ public class ReservationServiceImplTest {
     private RoomRepositoryService roomRepositoryService;
 
     @Mock
-    private EmailSenderService emailSenderService;
-
-    @Mock
-    private MessageSource messageSource;
+    private ReservationConfirmationEmailService reservationConfirmationEmailService;
 
     @Test
     public void testGetReservationsOfUserShouldReturnListOfReservationModelsBelongingToUser(){
